Introducing views and blueprints
