+++ app/__init__.py
@@ -2,12 +2,21 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from config import Config
+from app.views import bp
+
+db = SQLAlchemy()
+migrate = Migrate()
 
 
-app = Flask(__name__)
-app.config.from_object(Config)
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
+def create_app(config_class=Config):
+    app = Flask(__name__)
+    app.config.from_object(config_class)
+    db.init_app(app)
+    migrate.init_app(app, db)
+
+    app.register_blueprint(bp)
+
+    return app
 
 
-from app import routes, models
+from app import models
