+++ run.py
@@ -1,5 +1,8 @@
-from app import app, db
+import app
+from app import db
 from app.models import Translation
+
+app = app.create_app()
 
 
 @app.shell_context_processor
